requirements.txt
pandas
numpy
scipy
scikit-learn
matplotlib
seaborn
plotly
jupyter
ipywidgets
nbformat
wordcloud
nltk
textblob
streamlit
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-multipart
scikit-learn
category_encoders
textblob
nltk